fixed more conflicts
